+++ GameAssembly/GClass14.cs
@@ -32,14 +32,6 @@ public class GClass14 : GInterface0
 					{
 						GClass36.smethod_0("LOI NHAN  MESS THU 1");
 					}
-					try
-					{
-						Thread.Sleep(5);
-					}
-					catch (Exception)
-					{
-						GClass36.smethod_0("LOI NHAN  MESS THU 2");
-					}
 				}
 			}
 			catch (Exception ex3)
@@ -269,6 +261,30 @@ public class GClass14 : GInterface0
 		return bool_1 && tcpClient_0 != null && binaryReader_0 != null;
 	}
 
+	public static int GetSendQueueCount()
+	{
+		try
+		{
+			return gclass15_0.list_0.Count;
+		}
+		catch
+		{
+			return 0;
+		}
+	}
+
+	public static int GetReceiveQueueCount()
+	{
+		try
+		{
+			return gclass88_0.method_2();
+		}
+		catch
+		{
+			return 0;
+		}
+	}
+
 	public void setHandler(GInterface3 msgHandler)
 	{
 		ginterface3_0 = msgHandler;
@@ -319,6 +335,7 @@ public class GClass14 : GInterface0
 		if (GClass172.smethod_0().bool_1 && GClass151.bool_1)
 		{
 			tcpClient_0 = GClass147.smethod_0(host, port, GClass172.string_8, GClass172.smethod_0().int_1, GClass172.string_6, GClass172.string_7);
+			tcpClient_0.NoDelay = true;
 			networkStream_0 = tcpClient_0.GetStream();
 			binaryReader_0 = new BinaryReader(networkStream_0, new UTF8Encoding());
 			binaryWriter_0 = new BinaryWriter(networkStream_0, new UTF8Encoding());
@@ -336,6 +353,7 @@ public class GClass14 : GInterface0
 		{
 			tcpClient_0 = new TcpClient();
 			tcpClient_0.Connect(host, port);
+			tcpClient_0.NoDelay = true;
 			networkStream_0 = tcpClient_0.GetStream();
 			binaryReader_0 = new BinaryReader(networkStream_0, new UTF8Encoding());
 			binaryWriter_0 = new BinaryWriter(networkStream_0, new UTF8Encoding());